+++ include/roq/format_str.hpp
@@ -8,30 +8,29 @@
 #include <source_location>
 #include <string_view>
 
-// note!
-// this class captures the source location + does compile-time format string checking (c++20, only)
-// strongly inspired by the discussion here: https://github.com/fmtlib/fmt/issues/2022
+// capture source location
 
 namespace roq {
 
 namespace detail {
-// may truncate to N
+// compile-time evaluated string
 template <std::size_t N>
 struct static_string final {
-  consteval static_string(std::string_view const &sv) : length_{std::min(N, std::size(sv))}, buffer_{create(sv, length_)} {}
+  consteval static_string(std::string_view const &str) : length_{std::min(N, std::size(str))}, buffer_{create(str, length_)} {}
 
-  static_string(static_string const &) = default;
   static_string(static_string &&) = delete;
+  static_string(static_string const &) = default;
 
   constexpr operator std::string_view() const { return {std::data(buffer_), length_}; }
 
  protected:
-  static constexpr auto create(std::string_view const &sv, std::size_t length) {
+  static consteval auto create(std::string_view const &str, std::size_t length) {
+    static_assert(N > 0);
     std::array<char, N> buffer;
     for (std::size_t i = 0; i < length; ++i)
-      buffer[i] = sv[i];
+      buffer[i] = str[i];
     for (std::size_t i = length; i < N; ++i)
-      buffer[i] = static_cast<char>(0xEF);  // debug
+      buffer[i] = '\0';
     return buffer;
   }
 
@@ -39,27 +38,16 @@ struct static_string final {
   std::size_t const length_;
   std::array<char, N> const buffer_;
 };
-
-// note! like fmt::detail::check_format_string, but constexpr
-template <typename... Args>
-constexpr std::string_view check_format_string(std::string_view const &str) {
-  if constexpr (sizeof...(Args) > 0) {
-    auto s = fmt::basic_string_view<char>(str);
-    using checker = fmt::detail::format_string_checker<char, fmt::remove_cvref_t<Args>...>;
-    fmt::detail::parse_format_string<true>(s, checker(s));
-  }
-  return str;
-}
 }  // namespace detail
 
-template <typename... Args>
-struct basic_format_str final {
-  using file_name_type = detail::static_string<32>;
+struct format_str final {
   template <typename T>
-  consteval basic_format_str(T const &str, std::source_location const loc = std::source_location::current())
-      : str{detail::check_format_string<Args...>(str)}, file_name{extract_basename(loc.file_name())}, line{loc.line()} {}
+  consteval format_str(T const &str, std::source_location const loc = std::source_location::current())
+      : str{str}, file_name{extract_basename(loc.file_name())}, line{loc.line()} {}
 
-  fmt::string_view const str;
+  using file_name_type = detail::static_string<32>;  // note! string could be truncated
+
+  std::string_view const str;
   file_name_type const file_name;
   std::uint32_t const line;
 
@@ -75,11 +63,7 @@ struct basic_format_str final {
       return tmp;
     return tmp.substr(++pos);
   }
-};  // namespace roq
-
-template <typename... Args>
-using format_str = basic_format_str<fmt::type_identity_t<Args>...>;
-
+};
 }  // namespace roq
 
 template <size_t N>