+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Added
+
+* Support MassQuote (#483)
+
 ## 1.0.4 &ndash; 2024-12-30
 
 ## 1.0.3 &ndash; 2024-11-26