+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes will be documented in this file.
 
 ## Head
 
+## 1.0.5 &ndash; 2025-03-26
+
 ### Changed
 
 * Upgrade the `format_str` helper class to work with `fmt 11.1` (#488)